Delete : idea 파일 제거
