--- .idea/codeStyles
@@ -1,119 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="ProjectCodeStyleConfiguration">
-    <code_scheme name="Project" version="173">
-      <codeStyleSettings language="XML">
-        <indentOptions>
-          <option name="CONTINUATION_INDENT_SIZE" value="4" />
-        </indentOptions>
-        <arrangement>
-          <rules>
-            <section>
-              <rule>
-                <match>
-                  <AND>
-                    <NAME>xmlns:android</NAME>
-                    <XML_ATTRIBUTE />
-                    <XML_NAMESPACE>^$</XML_NAMESPACE>
-                  </AND>
-                </match>
-              </rule>
-            </section>
-            <section>
-              <rule>
-                <match>
-                  <AND>
-                    <NAME>xmlns:.*</NAME>
-                    <XML_ATTRIBUTE />
-                    <XML_NAMESPACE>^$</XML_NAMESPACE>
-                  </AND>
-                </match>
-                <order>BY_NAME</order>
-              </rule>
-            </section>
-            <section>
-              <rule>
-                <match>
-                  <AND>
-                    <NAME>.*:id</NAME>
-                    <XML_ATTRIBUTE />
-                    <XML_NAMESPACE>http://schemas.android.com/apk/res/android</XML_NAMESPACE>
-                  </AND>
-                </match>
-              </rule>
-            </section>
-            <section>
-              <rule>
-                <match>
-                  <AND>
-                    <NAME>.*:name</NAME>
-                    <XML_ATTRIBUTE />
-                    <XML_NAMESPACE>http://schemas.android.com/apk/res/android</XML_NAMESPACE>
-                  </AND>
-                </match>
-              </rule>
-            </section>
-            <section>
-              <rule>
-                <match>
-                  <AND>
-                    <NAME>name</NAME>
-                    <XML_ATTRIBUTE />
-                    <XML_NAMESPACE>^$</XML_NAMESPACE>
-                  </AND>
-                </match>
-              </rule>
-            </section>
-            <section>
-              <rule>
-                <match>
-                  <AND>
-                    <NAME>style</NAME>
-                    <XML_ATTRIBUTE />
-                    <XML_NAMESPACE>^$</XML_NAMESPACE>
-                  </AND>
-                </match>
-              </rule>
-            </section>
-            <section>
-              <rule>
-                <match>
-                  <AND>
-                    <NAME>.*</NAME>
-                    <XML_ATTRIBUTE />
-                    <XML_NAMESPACE>^$</XML_NAMESPACE>
-                  </AND>
-                </match>
-                <order>BY_NAME</order>
-              </rule>
-            </section>
-            <section>
-              <rule>
-                <match>
-                  <AND>
-                    <NAME>.*</NAME>
-                    <XML_ATTRIBUTE />
-                    <XML_NAMESPACE>http://schemas.android.com/apk/res/android</XML_NAMESPACE>
-                  </AND>
-                </match>
-                <order>ANDROID_ATTRIBUTE_ORDER</order>
-              </rule>
-            </section>
-            <section>
-              <rule>
-                <match>
-                  <AND>
-                    <NAME>.*</NAME>
-                    <XML_ATTRIBUTE />
-                    <XML_NAMESPACE>.*</XML_NAMESPACE>
-                  </AND>
-                </match>
-                <order>BY_NAME</order>
-              </rule>
-            </section>
-          </rules>
-        </arrangement>
-      </codeStyleSettings>
-    </code_scheme>
-  </component>
-</project>
